requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
pillow
